front-sick-lidar-measurement/requirements.txt
requests==2.32.3
streamlit==1.37.1
open3d==0.18.0
pillow==10.4.0